record_batcher: Move from transform/logging to kafka/client
Useful in audit_log_manager as well as transform logging
