+++ src/v/kafka/client/CMakeLists.txt
@@ -1,4 +1,16 @@
 v_cc_library(
+  NAME kc_record_batcher
+  HDRS
+    record_batcher.h
+  SRCS
+    record_batcher.cc
+  DEPS
+    v::config
+    v::model
+    v::storage
+)
+
+v_cc_library(
   NAME kafka_client
   SRCS
     assignment_plans.cc
