+++ src/v/kafka/client/test/BUILD
@@ -1,9 +1,15 @@
-load("//bazel:test.bzl", "redpanda_cc_btest", "redpanda_test_cc_library")
+load("//bazel:test.bzl", "redpanda_cc_btest", "redpanda_cc_gtest", "redpanda_test_cc_library")
 
 redpanda_test_cc_library(
     name = "utils",
+    srcs = [
+        "utils.cc",
+    ],
     hdrs = [
         "utils.h",
+    ],
+    implementation_deps = [
+        "//src/v/random:generators",
     ],
     include_prefix = "kafka/client/test",
     deps = [
@@ -98,5 +104,20 @@ redpanda_cc_btest(
         "@boost//:test",
         "@seastar",
         "@seastar//:testing",
+    ],
+)
+
+redpanda_cc_gtest(
+    name = "record_batcher_test",
+    timeout = "short",
+    srcs = ["record_batcher_test.cc"],
+    cpu = 1,
+    memory = "128MiB",
+    deps = [
+        ":utils",
+        "//src/v/base",
+        "//src/v/kafka/client:record_batcher",
+        "//src/v/test_utils:gtest",
+        "@googletest//:gtest",
     ],
 )
