+++ src/v/kafka/client/test/CMakeLists.txt
@@ -22,6 +22,29 @@ rp_test(
   LABELS kafka
 )
 
+v_cc_library(
+  NAME kc_test_utils
+  HDRS
+    utils.h
+  SRCS
+    utils.cc
+  DEPS
+    v::utils
+)
+
+rp_test(
+  UNIT_TEST
+  GTEST
+  BINARY_NAME test_kc_record_batcher
+  SOURCES
+    record_batcher_test.cc
+  LIBRARIES
+    v::gtest_main
+    v::kc_record_batcher
+    v::kc_test_utils
+  LABELS kafka
+)
+
 rp_test(
   FIXTURE_TEST
   BINARY_NAME kafka_client
