+++ src/v/transform/logging/rpc_client.cc
@@ -13,9 +13,9 @@
 
 #include "cluster/metadata_cache.h"
 #include "hashing/murmur.h"
+#include "kafka/client/record_batcher.h"
 #include "logger.h"
 #include "model/namespace.h"
-#include "record_batcher.h"
 
 #include <seastar/core/future.hh>
 #include <seastar/coroutine/as_future.hh>
@@ -41,7 +41,7 @@ rpc_client::write(model::partition_id pid, io::json_batches batches) {
     auto max_batch_size = cfg->properties.batch_max_bytes.value_or(
       config::shard_local_cfg().kafka_batch_max_bytes());
 
-    record_batcher batcher{max_batch_size};
+    kafka::client::record_batcher batcher{max_batch_size};
 
     while (!batches.empty()) {
         auto json_batch = std::move(batches.front());
