+++ src/v/transform/logging/tests/BUILD
@@ -31,21 +31,6 @@ redpanda_cc_gtest(
 )
 
 redpanda_cc_gtest(
-    name = "record_batcher_test",
-    timeout = "short",
-    srcs = ["record_batcher_test.cc"],
-    cpu = 1,
-    memory = "128MiB",
-    deps = [
-        ":utils",
-        "//src/v/base",
-        "//src/v/test_utils:gtest",
-        "//src/v/transform/logging:record_batcher",
-        "@googletest//:gtest",
-    ],
-)
-
-redpanda_cc_gtest(
     name = "log_manager_test",
     timeout = "short",
     srcs = ["log_manager_test.cc"],
